Scaling

+++ Assets/Scripts/AudioManipulation.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 
 public class AudioManipulation : MonoBehaviour
 {
-    public static void BitCrusher(AudioClip original, int compression)
+    public static void BitCrusher(AudioClip original, float compression)
     {
         // Acts as a fake audio compressor. Compression is the power of 2 Hz.
         float[] samples = new float[original.samples * original.channels];
